# Path B: Performance Optimization for NanoHFT