… examples/advanced/nanohft/.context/DfLSS/01_project_charter.md
… examples/advanced/nanohft/.context/DfLSS/02_voice_of_customer.md
… examples/advanced/nanohft/.context/DfLSS/03_critical_to_quality.md
… examples/advanced/nanohft/.context/diagrams/architecture.md
… examples/advanced/nanohft/.context/docs.md
… examples/advanced/nanohft/.context/index.md
… examples/advanced/nanohft/JOURNEY_SUMMARY.md
… examples/advanced/nanohft/PATH_B_OPTIMIZATION.md
… examples/advanced/nanohft/README.md
… examples/advanced/nanohft/core/legacy_strategy_temp.py
… examples/advanced/nanohft/core/strategy.py
… examples/advanced/nanohft/cython_ext/signals_fast.pyx
… examples/advanced/nanohft/demo_backtest.py
… examples/advanced/nanohft/profile_simple.py
… examples/advanced/nanohft/run_backtest.py
… examples/advanced/nanohft/tests/test_risk_guards.py
… examples/advanced/nanohft/tests/test_signals.py